Don't wait for DMA with wait_for_finish_blocking
printf() calls seems to mess things up while DMA is happening. There was
also a race condition where control returns to main at the same time
IMU_Finish_Burst() is called, which may have caused problems. This way
we just wait for a boolean to be set by the callback.